Continue working on push notifications
