+++ frontend/src/hooks/useAuth.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import firebase from "firebase";
+import { firebase } from "../firebase";
 
 function useAuth() {
     const [user, setUser] = useState(null);
